pyproject.toml
[project]
name = "obsidian-export"
version = "0.1.0"
description = "Convert Obsidian-flavored Markdown to PDF and DOCX via a 5-stage pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Matthias Christenson" },
]
keywords = ["obsidian", "markdown", "pdf", "docx", "pandoc", "export", "converter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business",
]
dependencies = [
    "pyyaml>=6.0,<7",
    "click>=8.0,<9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9", "hypothesis>=6.0,<7", "pytest-cov>=5.0,<6"]

[project.urls]
Repository = "https://github.com/neuralsignal/obsidian-export"

[project.scripts]
obsidian-export = "obsidian_export.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["obsidian_export"]

[tool.hatch.build.targets.sdist]
include = ["obsidian_export/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM108"]

[tool.pytest.ini_options]
testpaths = ["tests"]
